requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --index-url=https://artefact-repository/repository/pypi/simple
#
--index-url https://artefact-repository/repository/pypi/simple

babel==2.12.1
    # via flask-babel
blinker==1.6.2
    # via flask
brotli==1.1.0
    # via flask-compress
cachelib==0.9.0
    # via flask-session2
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   misaka
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via flask
cryptography==41.0.4
    # via -r requirements.in
deepmerge==1.1.0
    # via govuk-frontend-wtf
defusedxml==0.7.1
    # via -r requirements.in
dicttoxml==1.7.16
    # via -r requirements.in
dnspython==2.4.2
    # via email-validator
email-validator==2.0.0.post2
    # via -r requirements.in
flask==2.3.3
    # via
    #   -r requirements.in
    #   flask-babel
    #   flask-compress
    #   flask-script
    #   flask-session2
    #   flask-wtf
    #   govuk-frontend-wtf
    #   landregistry-enhanced-logging
    #   landregistry-exception-handlers
    #   landregistry-healthchecks
    #   landregistry-security-headers
    #   landregistry-trace-id
flask-babel==3.1.0
    # via -r requirements.in
flask-compress==1.14
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-session2==1.3.1
    # via -r requirements.in
flask-wtf==1.1.1
    # via
    #   -r requirements.in
    #   govuk-frontend-wtf
geographiclib==2.0
    # via geopy
geopy==2.4.0
    # via -r requirements.in
govuk-frontend-jinja==2.7.0
    # via
    #   -r requirements.in
    #   govuk-frontend-wtf
govuk-frontend-wtf==2.5.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
hmlr-frontend-jinja==1.0.0
    # via -r requirements.in
idna==3.4
    # via
    #   email-validator
    #   requests
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via
    #   flask
    #   flask-babel
    #   govuk-frontend-jinja
    #   govuk-frontend-wtf
    #   hmlr-frontend-jinja
jwt-validation @ git+http://internal-git-host/llc-beta/jwt-validation.git@v1.2.3
    # via -r requirements.in
landregistry-enhanced-logging[flask]==2.0.2
    # via -r requirements.in
landregistry-exception-handlers==2.0.1
    # via -r requirements.in
landregistry-healthchecks==3.0.0
    # via -r requirements.in
landregistry-security-headers==1.0.3.1
    # via -r requirements.in
landregistry-trace-id==2.0.1
    # via landregistry-enhanced-logging
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
marshmallow==3.20.1
    # via
    #   -r requirements.in
    #   jwt-validation
misaka==2.1.1
    # via -r requirements.in
namedlist==1.8
    # via -r requirements.in
numpy==1.26.0
    # via shapely
packaging==23.1
    # via
    #   gunicorn
    #   marshmallow
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2022.7.1
    # via
    #   flask-babel
    #   flask-session2
redis==5.0.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   jwt-validation
    #   landregistry-healthchecks
shapely==2.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
urllib3==2.0.5
    # via requests
werkzeug==2.3.7
    # via
    #   -r requirements.in
    #   flask
    #   landregistry-exception-handlers
wtforms==3.0.1
    # via
    #   flask-wtf
    #   govuk-frontend-wtf
